test: extract the resume neutrality harness
The neutrality test carried its scenario and the machinery around it in one
class: building the world from a configuration, registering plugins, serializing
through the real encoder, restoring, and describing the full state for
comparison. A second scenario would either have duplicated that or been squeezed
into the same run.

The machinery moves to ResumeNeutralityHarness; what stays behind is the
scenario. The plugin list moves with it, with the mutation rate as a parameter,
so both scenarios configure the same plugins and PluginCoverageTest keeps
checking exactly one list.

No behaviour change: the four existing scenarios still pass unmodified.
